requirements.txt
flask==2.0.1
werkzeug==2.0.1
sqlalchemy==1.4.23
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-jwt-extended==4.2.3
flask-cors==3.0.10
pymysql==1.0.2
cryptography==38.0.4
python-dotenv==0.21.0
alembic==1.7.1
markupsafe==2.0.1
itsdangerous==2.0.1
jinja2==3.0.1
